Proper error variant for non exhaustive match (#1772)
Argument to match expressions that wouldn't match any of the branches
used to have a pretty bad error message ("expecting Enum, found Enum").
This commit creates a proper variant for non-matching expressions,
pointing to the match expression and to the value the argument evaluated
to.

Currently, one position that might be lost in translation is where the
match expression was applied. It's a bit harder to track; for now, this
commit is a trivial change and is a strict improvement over the previous
situation. Let's reconsider later if this is insufficient.