Fix expressions evaluated multiple times
This commit fixes an issue that has been introduced by the removal of
generated variables, together with the `should_update` optimization,
which tries to avoid thunk update that are unnecessary.

However, as detailed in the added documentation for should_update, this
didn't work well with the removal of generated variables, which would
materialize by not sharing inner expressions for variables whose value
was already a data structure in weak head normal form (arrays, records
or now enum variants). Inner expressions would be recomputed.

This commit fixes the issue by adding a condition that datastructure
that aren't yet closurized shouldn't eschew thunk update.

Doing so, this commit also adds the missing closurization of ADTs, which
caused enum variant's arguments to also be potentially re-evaluated many
time.

Co-authored-by: Viktor Kleen <viktor.kleen@tweag.io>