Improve cache invalidation for failed imports.
We were already tracking import dependencies for cache invalidation,
but only for imports that exist. Here, we add support for imports that
fail because of I/O errors: we remember the filename of the failed import,
and whenever nls becomes aware of a new file with that name, it
invalidates the importers.