Introduce a proper trait for to_type
The `to_type` helper to convert a unification type to a proper type that
can be displayed in error messages was only exposed for types, although
it defined internally other variants for e.g. enum rows and record rows.

This commit refactors the code to put this functionality in a proper
`ToType` trait, such that `to_type` now works for other kind of types as
well and can be used to make the definitions of `EnumRowConflict` and
`RecordRowConflict` more precise (use record rows and enum rows directly
instead of wrapping them in a top-level `Type`).