Fix error reporting of row conflict
The error reporting of row conflict was off, sometimes mixing different
types involved in the issue, and ignoring the expected and inferred
types from the parent error.

This commit eliminates some data from those errors, mostly because they
can't always be known at the error raising site (or at least not without
significant clones of other types), and don't seem to be reliable. We
stay conservative for now, and just store the conflicting row together
with the original types that were involved in the parent unification.

We can improve from there if practical situations show that this is
insufficient.