[Fiix & Refactor] Row conflict error (#1808)
* Refactor: use named params for typecheck errors

Use named parameters (anonymous structs) for
`nickel_lang_core::error::TypecheckError`, instead of relying on the
position of parameters in the variant.

Also split the `RowConflict` error in two different ones, to reflect the
recent refactoring of the internal typechecker errors.

* Introduce a proper trait for to_type

The `to_type` helper to convert a unification type to a proper type that
can be displayed in error messages was only exposed for types, although
it defined internally other variants for e.g. enum rows and record rows.

This commit refactors the code to put this functionality in a proper
`ToType` trait, such that `to_type` now works for other kind of types as
well and can be used to make the definitions of `EnumRowConflict` and
`RecordRowConflict` more precise (use record rows and enum rows directly
instead of wrapping them in a top-level `Type`).

* Fix error reporting of row conflict

The error reporting of row conflict was off, sometimes mixing different
types involved in the issue, and ignoring the expected and inferred
types from the parent error.

This commit eliminates some data from those errors, mostly because they
can't always be known at the error raising site (or at least not without
significant clones of other types), and don't seem to be reliable. We
stay conservative for now, and just store the conflicting row together
with the original types that were involved in the parent unification.

We can improve from there if practical situations show that this is
insufficient.

* Rename error field mismatch -> cause

* Update tests to latest error renamings

* Fix and add snapshot test for row conflicts

* Fix tests annotations after renaming