Fix unification order
A test revelead that the typechecking errors were off sometimes for
destructuring lets: if the destructured value didn't correspond to the
pattern, the type elaborated from the pattern appeared as the inferred
type of the expression, and the inferred type of the expression as the
expected type (i.e. the two were reversed).

This causes incorrect error messages like "extra row" instead of
"missing row". This commits fix the issue by unifying the type deduced
from the pattern and the type inferred from the bound expression in the
right order (unification is symmetric, as far as typechecking is
concerned, but for error reporting we do distinguish between expected
and inferred type).