ADT destructuring (#1812)
* Add unwrap_enum_tag primop

Add a primitive operation to extract the argument from an enum variant,
which will prove useful to enum matching and destructuring.

* Add get_tag and is_variant enum primops

* ADTs destructuring

This commit adds support for enum variant patterns in destructuring
forms (let bindings and function declarations). This paves the way for
matching on ADTs once we allow patterns not only appear in a
destructuring binding, but also in a match expression.

* Make get_tag work on simple enum tags as well

* Fix pretty printing of enum_unwrap_variant

* Additional tests for ADTs destructuring

* Fix unification order

A test revelead that the typechecking errors were off sometimes for
destructuring lets: if the destructured value didn't correspond to the
pattern, the type elaborated from the pattern appeared as the inferred
type of the expression, and the inferred type of the expression as the
expected type (i.e. the two were reversed).

This causes incorrect error messages like "extra row" instead of
"missing row". This commits fix the issue by unifying the type deduced
from the pattern and the type inferred from the bound expression in the
right order (unification is symmetric, as far as typechecking is
concerned, but for error reporting we do distinguish between expected
and inferred type).

* Fix type of is_variant to not break parametricity

* Remove useless TODO: tests will catch this

* Format stdlib internals