Add %field_is_defined%
To safely check if a pattern matches a value (safely meaning without
throwing an uncatchable error), we need a primitive operation to check
if we can safely access the field of a record. To do so, we need the
field to exists _and_ to have a definition. This is what the added
`field_is_defined` checks.