Correctly handle record pattern tail
The partial implementation of record pattern compilation didn't handle
the record pattern tail yet: for example, a closed pattern `{foo, bar}`
would match a larger record `{foo=1, bar=2, baz=3}`.

This commit additionally keeps the current rest of the record being
matched, that is the original record minus all the matched fields, to
either check that it is empty at the end (closed pattern), or to bind
this rest to an identifier if the pattern captures the rest with `{foo,
bar, ..rest}`.