Switch from Span to TermPos for patterns (#1819)
Patterns were annotated with a span, which must always be a well-defined
location in the source file. The implementation of full pattern matching
- i.e. allowing match expression to have arbirary patterns on the left
hand side of "=>" - showed that this is an issue, because the enum
contract is implemented using match, and thus generate a match
expression without a definite position.

This commit switches to optional position instead, delaying the fearful
`pos.unwrap()` to the latest moment possible, when we need to generate a
contract and build a `Label` during a pattern destructuring, which
shouldn't affect the contract-generated match expressions.