Add InvalidContractError (#1824)
This commit adds a specific parse error which detects upfront
non-sensical contract annotations. More precisely, when a bare constant
(null, boolean, string, number) or a bare array is used as a contract in
a type expression or in an annotation. This will hopefully catch common
typo like using `:` instead of `=` when coming from JSON-like languages