Fix the semantics of default values in patterns (#1826)
* Fix semantics of default values in patterns

The previous semantics of default value in patterns (set with the `?`
annotation) were kinda accidental: at the time they were introduced, it
was much easier to implement default values by the `default` merge
priority. However, it's not aligned with intuition when using default
value in match statement or function argument patterns: the goal here is
mostly to say, if the field isn't defined, I want to use fallback value.
The previous implementation, for example, could fail with merge conflict
or happily merge two default values. In fact, we want to give the
priority to any value that exists in the matched value, whatever is its
actual merge priority.

This commit implements the desired semantics: pattern default values are
only set if the field isn't there, or is there but has no definition. In
that case, we rely on merging to avoid erasing any metadata present
previously (for example, if we set a default value in `{foo | String}`,
we don't want to drop the `String` contract, so we can't use
std.record.update blindly).

* Add From<_> instances to ease Field creation