Fix unbound identifier when querying in REPL
Following a refactoring that landed in 1.4, the current environment of
a REPL session wasn't taken into account when querying an expression.
This mean that any value introduced by a toplevel let during the session
would evaluate fine but raise an unbound identifier error if one tries
to query it with `:query some_id`.

This commit fixes this by correctly propagating the environment in the
subfunctions involved in querying.