Specialized pattern compilation for enum tags (#1846)
The recent introduction of full pattern matching incurred a performance
regression for very large enum contracts.

Indeed, previously enum-only pattern matching would directly index into
a hashmap of the branches of the pattern - where the key is the enum tag
which is the pattern associated with the branch -, giving amortized
constant time.

The more powerful and generic pattern matching introduced later instead
compiles each pattern to a (potentially complex) booleanc check. This is
necessary in general, but for match statements that are just composed of
bare enum tags, this trade a constant time operation for a linear (in
the number of branches). This shown important regressions on some code
bases making heavy usage of very large enum types/contracts.

This commit specializes the compilation of match expressions when all
the patterns are enum tags to recover the old behavior. We just reuse
the unary operator `%match%`, which was precisely used to implement the
old pattern matching, and was just standing there unused since the
introduction of the new match expression compilation.