1.5.0 release (#1853)
* [release.sh] update to 1.5.0

* Fix LSP not compiling when format feat is disabled

* Add 1.5 release notes

* Don't mention current release in the manual intro

We used to mention the current release version in the manual intro, but
this is tedious to keep in sync, and not very valuable - this
information is really easy to get by many other mean. The original
intent was to insist that Nickel was past 1.0, and thus was considered
relatively stable and usable in production. The text has been reworded
to say that instead.