Fix polymorphic field type error
Using a polymorphic type annotation on a record field, as in
`{id : forall a. a -> a = std.function.id}` would unexpectedly file with
a failure to unify `forall a. a -> a` with `a -> a`, which indicates
that the type annotation wasn't properly instantiated before being used
to check the value of the field.

It turns out the typechecking of the field per se is fine, but the issue
comes from the recursive environment: in a recursive record, we need to
build an environment with types for all record fields _before_ we can
start to actually typecheck each field. When the field isn't annotated,
we use a fresh unification variable. When typechecking the corresponding
field, we unify what's in the type environment (the unification
variable, but which might have been unified with another type at this
point) with the type inferred for the field.

However, when the field has a type annotation, we use this annotation
for the recursive environment instead (which is important to keep
polymorphic types...polymorphic). In this case, we would unify a
polymorphic type with the inferred type, which is the place that was
missing the instantiation.

One possibility is to just add the missing instantiation, but in fact
this unification is useless when there is a type annotation: to avoid
this useless work, we rather remember which fields are annotated, and we
entirely skip the unification for them.