Fix missing newline in markdown output (#1880)
A recent commit fixed the `nickel doc` markdown output having spurious
backslashes introduced. However, it also re-introduced a previous issue
- which was why the problematic hard line breaks were introduced in the
first place - which is that the list of types and contracts of a field
and the custom documentation might be squashed together because the
renderer wouldn't properly insert a newline after the generated list.

After some investigation, it seems to be a bug in the comrak renderer.
However, this bug doesn't show when the markdown is parsed from an
actual source (instead of being programmatically generated). It turns
out comrak correctly inserting this required newline relies on the
unspoken assumption that every markdown list item content is wrapped as
a paragraph. This commit thus fixes the issue by inserting the missing
paragraph wrapper.