Update to latest Topiary and tree-sitter-nickel
For some reason, the version of Topiary and of the Nickel tree sitter
grammar were lagging behind, before the support for pattern matching was
added, which means current master wasn't able to format a file with
pattersn in it - maybe it was a rollback because of the Topiary install
issue (https://github.com/tweag/topiary/issues/690).

The former issue has been fixed, so this commit update to latest Topiary
in order to get formatting of patterns.