Include field path in non serializable error (#1905)
When failing to serialize a value, the error sometimes doesn't point to
a usable location (that is, it points to the definition site of some
function somewhere, but it's impossible to know where this function was
actually used, which is usually where the issue is).

This commit enriches the error message of non-serializable errors with a
path, à la JSON pointer (such that it also integrates array indices),
whenever possible. This should make debugging non serializable error
much nicer.