Add span information for TOML imports
This commits use a feature of the toml crate allowing to embed span
information into the deserialized data (imported from Nickel). This
makes it possible to give nice error messages and point to within a TOML
file when the application of a Nickel contract to an existing TOML
configuration fails, or in case of unmergeable values.

Doing so requires a bit of boilerplate and brings an additional crate in
(although small). It can also impact deserialization performance,
because we need to first deserialize in an intermediate TOML-like
datastructure. For those reasons, spanned deserialization has been gated
by the `spanned-deser` feature, which is enabled by default, but can be
disabled if this isn't useful.