use Topiary's published crates over git (#1919)
Now that Topiary is published and crates.io (and tree-sitter-nickel as
well), we move away from git revision and use proper public versions for
those. Doing so, we can also crap the special casing that was done
during release to disable those dependencies and we'll now be able to
release on crates.io a version with formatting enabled.

Co-authored-by: Yann Hamdaoui <yann.hamdaoui@tweag.io>