Update the release.sh script (#1952)
Update the release script by adding some missing `git add`, which were
lost when switching to using Topiary from crates.io.

Doing so, we also remove the `--dry-run` step before publishing: the
usual workflow is to perform the dry run test, and if it succeeds,
immediately do the actual publication. This isn't very useful: if the
dry run fails, we could have run the actual publication as well, which
would have fail in the same way, so this makes no difference. However,
if the dry run succeeds, we recompile each package one more time from
scratch, performing the exact same work for publication, which is
wasteful. At this point of the release script, we clearly intend to
release and we should just use `cargo publish` normally.