Don't blacklist files forever in the lsp
When the evaluator in the LSP times out or overflows its stack, the
faulty file is blacklisted so that it doesn't get evaluated any more.
This blacklist used to be permanent, which is generally not desirable
since it makes the LSP unusable.

Change that to only blacklist the file for a fixed delay currently
(hardcoded at 30s).