Additional tests for relaxed function comparison
Check the relaxed function comparison, which errors out when trying to
compare two functions, but can compare a function to a value of a
different type (by always returning false).

Tests that functions can't be compared were already there, but this
commit adds tests that function can be compared to other values.